requirements.txt
# Python Version
# Requires Python 3.12.9
# ----------------------

# Critical dependencies - install first
pandas==2.2.3
numpy==2.2.4

# Core dependencies
python-dotenv==1.1.0

# API integrations
openai==1.72.0
google-generativeai==0.8.4
flask==2.0.1
flask-cors==3.0.10
requests==2.32.3

# Web & deployment
gunicorn==20.1.0
werkzeug==2.0.3
flask-restful==0.3.10

# Data processing and visualization
pytz==2025.2
folium==0.19.5
markdown==3.8
pillow==11.1.0
branca==0.8.1
shapely==2.1.0
contourpy==1.3.1
reportlab==4.3.1
matplotlib==3.10.1

# Development and testing
ipython==8.12.0
pytest==8.3.5
pytest-mock==3.14.0

# Process management
psutil==7.0.0

# Additional dependencies found in your environment
jinja2==3.1.6
pydantic==2.11.3
pydantic-core==2.33.1
typing-extensions==4.13.2
certifi==2025.1.31
idna==3.10
urllib3==2.4.0
six==1.17.0
markupsafe==3.0.2
setuptools==75.8.0
wheel==0.45.0

# Note: Standard library modules don't need to be in requirements.txt:
# datetime, pathlib, pickle, hashlib, random, os, sys, json, re, time, math, 
# traceback, http.server, socketserver, threading, signal, platform, 
# webbrowser, shutil, subprocess 